fix types

+++ src/simulation.hpp
@@ -1224,7 +1224,7 @@ template <typename problem_t> void AMRSimulation<problem_t>::evolve()
 			break;
 		}
 
-		if (maxWalltime_ > 0 && getWalltime() > std::max(0.9 * maxWalltime_, maxWalltime_ - 300)) {
+		if (maxWalltime_ > 0 && getWalltime() > std::max(0.9 * maxWalltime_, static_cast<double>(maxWalltime_ - 300))) {
 			// we have exceeded the walltime limit
 			break;
 		}
